make a clean git mv instead of just renaming (2)
